+++ docker/docker-env.sh
@@ -5,11 +5,6 @@ pip3 install --upgrade pip
 pip3 install --requirement builder-requirements.txt
 PKG_VERSION=$(./version.sh)
 ENV_FN=$(./docker/docker-env-filename.sh)
-# export WHEELS=/app/cache/packages/wheels
-# HOST_CACHE_DIR=./cache/packages/$(./docker/docker-arch.sh)
-# mkdir -p "$HOST_CACHE_DIR"
-# touch "$HOST_CACHE_DIR/x"
-# export HOST_CACHE_DIR
 rm -f "$ENV_FN"
 # shellcheck disable=SC1091,SC2129
 cat <<EOF >>"$ENV_FN"
@@ -18,7 +13,6 @@ CODEX_BASE_VERSION=$(./docker/docker-version-codex-base.sh)
 EOF
 echo "CODEX_BUILDER_BASE_VERSION=$(./docker/docker-version-codex-builder-base.sh)" >>"$ENV_FN"
 echo "CODEX_DIST_BUILDER_VERSION=$(./docker/docker-version-codex-dist-builder.sh)" >>"$ENV_FN"
-echo "CODEX_BUILDER_FINAL_VERSION=$(./docker/docker-version-codex-builder-final.sh)" >>"$ENV_FN"
 cat <<EOF >>"$ENV_FN"
 CODEX_ARCH_VERSION=$(./docker/docker-version-codex-arch.sh)
 CODEX_WHEEL=codex-${PKG_VERSION}-py3-none-any.whl